+++ test/linearAlgebra/petsc.cpp
@@ -57,6 +57,15 @@ TEST_CASE("MatrixAssembly - Petsc")
             {"solver", std::string {"Petsc"}},
         }};
 
+        NeoN::Dictionary subDict;
+        subDict.insert("pc_type", std::string("bjacobi"));
+        subDict.insert("sub_pc_type", std::string("ilu"));
+        /*subDict.insert(
+            "log_view", std::string(":/home/michael/petsc/src/ksp/ksp/tutorials/test2.txt")
+        );
+    */
+        solverDict.insert("options", subDict);
+
         // Create solver
         auto solver = NeoN::la::Solver(exec, solverDict);
 
@@ -88,6 +97,11 @@ TEST_CASE("MatrixAssembly - Petsc")
             NeoN::Dictionary solverDict {{
                 {"solver", std::string {"Petsc"}},
             }};
+
+            NeoN::Dictionary subDict;
+            subDict.insert("pc_type", std::string("bjacobi"));
+            subDict.insert("sub_pc_type", std::string("ilu"));
+            solverDict.insert("options", subDict);
 
             // Create solver
             auto solver = NeoN::la::Solver(exec, solverDict);